Add files via upload
Tools and Technologies Used in the Project:
This project is going to be very beginner-friendly.
The working of this software is entirely POP( Procedural Oriented Programming). Basic knowledge about python functions is needed.
Datetime module of python.
Required Skill set to Build the Project:
Python
Visual Studio Code or any other code editor.